Early Battle Map Update
Added initial class designs for ships, crews and battle maps.

Ships - Customizable vehicles used to perform battle in the battle maps. Crews are required to pilot a ship.

Crews - Object that represents the crew of a ship. Crews will level up and be able to be cutomized for specific needs such as expanding crew sizes and learning skills that can be used to improve specific ship functionalities.

Battle Map - Load, calculate and draw battle maps which include the ships fighting on the battle map, tile information and calculation for user input. Currently only ship that occupy can move, but still needs improvements such as checking if a tile is already occupied.

Update some class layout  and namings to match with intended coding conventions that are being developed.