Battle Menu Update
Added a new battle menu to the code that allows a player to select from a list of options. This battle menu will appear to the right of the ship that is currently taking its action turn. Be advised this is an initial implementation of this battle menu and is most likely to be updated to be more robust in the future.

Additional changes include updates to the battle map class to allow changes of state such as determining which ship will be next to take its action turn and change states based on the new battle menu selections. Also able to cancel out on some of these selections such as canceling the move selection to return to the battle menu.